pyproject.toml
[tool.poetry]
name = "console"
version = "0.1.0"
description = ""
authors = ["Yaroslav <yar1k_k12@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
drf-yasg = "^1.21.7"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = {version = "^7.3.0", extras = ["phonenumbers"]}
django-allauth = "^0.61.1"
django-rest-passwordreset = "^1.4.0"
ruff = "^0.3.4"
python-slugify = "^8.0.4"
pillow = "^10.3.0"
gunicorn = "^21.2.0"
itsdangerous = "^2.1.2"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
